feat: initial commit
Supports a basic terraform workflow, where there is one root directory
with a structure like the following:

/terraform
  /project
    main.tf
    terragrunt.hcl
  /projects_with_common_module
    /_module
      main.tf
    /project
      terragrunt.hcl
  terragrunt.hcl

Dependencies between projects are also accounted for, but only if the
dependency path follows a very strict convention. They must always be
something like the following:

"${get_terragrunt_dir()}/../../projects_with_common_module/project"

and so on, for both module and module-less scenarios.

Additionally, there is a Dockerfile for planned usage from CI. It's only
been tested on podman, but should work elsewhere.